style: reformat the code for better readability | SCRUM-7
