+++ src/app/listFiles/layout.tsx
@@ -1,4 +1,4 @@
-import SideNav from "@/app/ui/sidenav";
+import SideNav from "@/app/ui/side-nav";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
