+++ src/app/ui/file-upload.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React, { useState, FormEvent, ChangeEvent } from "react";
-import { convertToTaipeiTime } from "@/app/lib/utils/dataUtils";
+import { convertToTaipeiTime } from "@/lib/utils/dataUtils";
 
 interface FileDataType {
   file_id: string;
