+++ src/app/ui/side-nav.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
 import Image from "next/image";
-import { HomeIcon } from "@heroicons/react/24/outline";
+// import { HomeIcon } from "@heroicons/react/24/outline";
 
 export default function SideNav() {
   return (
