+++ src/app/uploadFile/page.tsx
@@ -1,4 +1,3 @@
-import SendForm from "@/app/ui/send-form";
 import FileUpload from "@/app/ui/file-upload";
 
 export default function Page() {
